Bug 1366201 - P3. Update ffvp9/ffvp8 to 3.4 branch. r=gerald
Structure of code was slightly modified so that it should be no longer necessary to re-generate the config_*.h files, greatly simplifying the resync process.

MozReview-Commit-ID: Ap6HpJAANT6

UltraBlame original commit: d334969fa9f6d6148ecf60ef877d8b613de0da9a

+++ media/ffvpx/libavcodec/blockdsp.h
@@ -19,6 +19,7 @@
 #ifndef AVCODEC_BLOCKDSP_H
 #define AVCODEC_BLOCKDSP_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 #include "avcodec.h"
@@ -29,7 +30,7 @@
 
 
 typedef void (*op_fill_func)(uint8_t *block ,
-                             uint8_t value, int line_size, int h);
+                             uint8_t value, ptrdiff_t line_size, int h);
 
 typedef struct BlockDSPContext {
     void (*clear_block)(int16_t *block );
